codework/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-oauth-toolkit==2.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
drf-social-oauth2==1.3.0
ecdsa==0.18.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
oauthlib==3.2.2
pip==22.0.2
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.8
python-http-client==1.2.3
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sendgrid==2.2.1
setuptools==59.6.0
six==1.16.0
smtpapi==0.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
wrapt==1.15.0
